requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
